+++ grelmicro/task/__init__.py
@@ -1,6 +1,6 @@
 """Task."""
 
-from grelmicro.task._cron import FireInfo
+from grelmicro.task._cron import FireInfo, FireOutcome
 from grelmicro.task._tasks import Tasks
 from grelmicro.task.errors import (
     CronError,
@@ -13,6 +13,7 @@ from grelmicro.task.router import TaskRouter
 __all__ = [
     "CronError",
     "FireInfo",
+    "FireOutcome",
     "FunctionTypeError",
     "TaskAddOperationError",
     "TaskError",